+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["crates/*"]
+members = ["crates/*", "examples/*"]
 resolver = "2"
 
 [workspace.dependencies]
@@ -9,4 +9,5 @@ regex = "1.10"
 serde = { version = "1.0", features = ["derive"]}
 serde_json = "1.0"
 thiserror = "1.0"
+tokio = { version = "1.37", features = ["full"] }
 tracing = "0.1"
