+++ crates/ai/src/process/splitter/mod.rs
@@ -1,0 +1,1 @@
+pub mod simple;
