+++ crates/ai/src/process/splitter/simple.rs
