--- crates/ai/src/processors.rs
@@ -1,1 +1,0 @@
-pub mod simple_splitter;
